优化

+++ divine-modules/divine-mine/src/main/java/com/divine/mine/domain/dto/MineCarDto.java
@@ -62,7 +62,7 @@ public class MineCarDto extends BaseEntity {
     /**
      * 是否删除(0:未删除,1:已删除)
      */
-    private Long isDel;
+    private Integer isDel;
 
 
 }
