+++ divine-modules/divine-mine/src/main/java/com/divine/mine/domain/entity/MineQuality.java
@@ -51,6 +51,6 @@ public class MineQuality extends BaseEntity {
     /**
      * 是否删除(0:未删除,1:已删除)
      */
-    private Long isDel;
+    private Integer isDel;
 
 }
