+++ divine-modules/divine-mine/src/main/java/com/divine/mine/domain/vo/MineCarRefuelVo.java
@@ -87,7 +87,7 @@ public class MineCarRefuelVo implements Serializable {
      * 是否删除(0:未删除,1:已删除)
      */
     @ExcelProperty(value = "是否删除(0:未删除,1:已删除)")
-    private Long isDel;
+    private Integer isDel;
 
 
 }
